fix(sentrux): split sentrux_evolution.rs to clear this repo's own monolith gate
sentrux_evolution.rs had grown to 1330 lines / 36 functions after rustfmt
expansion, tripping this repo's own architecture/quality gate (loc>800, or
functions>25 with loc>400) — which sentrux-capability-gate CI enforces via
diagnosis.hospital's domain verdict. Confirmed by reproducing the exact CI
step locally: `code-intel run execute` reported outcome "domain_failed"
with diagnosis.hospital diagnostic "architecture gate failure" against the
single-file version, and "completed" after this split.

Splits what_if (rule hints, pollution/scope-candidate heuristics, the
8-scenario matrix, iso8601 timestamp) into a new sentrux_what_if.rs,
included as a private submodule of sentrux_evolution.rs and re-exported so
sentrux.rs's `sentrux_evolution::what_if(&repo)` call site is unchanged.
Shared helpers (as_array, evolution_bus_factor) marked pub(super). Each
file is now under 700 lines / under 20 top-level functions.

Also repins legacy/run-code-intel.ps1 and crates/code-intel-cli/src/
sentrux.rs + sentrux_analysis.rs in orchestration/internalization/*.json —
stale after the previous commit's edits (code-intel repin --write),
confirmed clean by cargo test -p code-intel --test declared_pins.

cargo test -p code-intel --bins --tests: 919 passed, 0 failed.
