+++ orchestration/internalization/rg.json
@@ -5,7 +5,7 @@
   "subject": { "name": "ripgrep inventory executable", "kind": "adapted_capability", "source": { "uri": "https://github.com/BurntSushi/ripgrep; installed-evidence=rg 15.1.0 (rev af60c2de9d)", "revision": "installed-version:15.1.0-af60c2de9d; local-native-source-sha256:9fad4644b8cfb88fbec6f4603e7fcb099dc296fdf29286a6909279fab1450d0e; local-conformance-sha256:1a19972c41c0c1183f549cba9601c6e5742208c6888cff223a8de7e587f26c53" }, "license": { "id": "MIT-OR-Unlicense-UPSTREAM-CLAIM-LOCAL-COPY-MISSING", "obligations": ["do not redistribute or upgrade from this record until the selected license text and package provenance are retained locally", "preserve exact inventory scope, exclusions, order normalization, snapshot lease, and failure semantics"] } },
   "adoption": { "rung": "invoke", "ownedBoundary": ["Pipeline owns inventory.rg request normalization, snapshot-controlled mirror, exclusions, artifact contract, and failure mapping", "ripgrep owns executable search and traversal behavior; this record does not authorize upgrade, vendoring, or reimplementation"], "necessityEvidence": { "evidenceIds": ["local:registry:inventory.rg:required", "local:r09:installed-rg-15.1.0-af60c2de9d", "gap:rg:package-provenance"], "checkedAt": 1783900800, "expiresAt": 1791676800 }, "compatibilityEvidence": { "evidenceIds": ["local:a00:inventory-parity", "local:r09:cross-platform-contract", "gap:rg:replacement-command-drill"], "checkedAt": 1783900800, "expiresAt": 1791676800 }, "conformanceEvidence": { "evidenceIds": ["local:r09:scope-exclusion-conformance", "local:r09:operation-trace", "gap:rg:representative-platform-matrix"], "checkedAt": 1783900800, "expiresAt": 1791676800 } },
   "operationTrace": [
-    { "integrationId": "inventory.rg", "operation": "run", "command": "legacy/run-code-intel.ps1 -RepoPath <repo-path> -Mode <mode>", "implementationIdentity": { "providerId": "ripgrep", "implementationId": "rg-15.1.0-af60c2de9d-via-compat-facade", "activation": "required production facade" }, "source": { "path": "legacy/run-code-intel.ps1", "sha256": "971120bad5d61b26e9d1c42ef68d760a3a75f28f69daf8863fd4973e08c4b5c0" }, "conformance": { "path": "crates/code-intel-cli/tests/capability_exec.rs", "sha256": "1a19972c41c0c1183f549cba9601c6e5742208c6888cff223a8de7e587f26c53", "testName": "normalized_inventory_matches_real_legacy_runner_with_custom_exclude" } },
+    { "integrationId": "inventory.rg", "operation": "run", "command": "legacy/run-code-intel.ps1 -RepoPath <repo-path> -Mode <mode>", "implementationIdentity": { "providerId": "ripgrep", "implementationId": "rg-15.1.0-af60c2de9d-via-compat-facade", "activation": "required production facade" }, "source": { "path": "legacy/run-code-intel.ps1", "sha256": "89778bb7f0cb6e470a9b1781df08710e78225c66827be36a168b840478507b7b" }, "conformance": { "path": "crates/code-intel-cli/tests/capability_exec.rs", "sha256": "1a19972c41c0c1183f549cba9601c6e5742208c6888cff223a8de7e587f26c53", "testName": "normalized_inventory_matches_real_legacy_runner_with_custom_exclude" } },
     { "integrationId": "inventory.rg", "operation": "capabilityExec", "command": "target/debug/code-intel.exe capability exec inventory.rg --request <request.json|-> --out <staging-dir>", "implementationIdentity": { "providerId": "ripgrep", "implementationId": "inventory.rg.compat+rg-15.1.0-af60c2de9d", "activation": "required production capability envelope" }, "source": { "path": "crates/code-intel-cli/src/capability_inventory.rs", "sha256": "9fad4644b8cfb88fbec6f4603e7fcb099dc296fdf29286a6909279fab1450d0e" }, "conformance": { "path": "crates/code-intel-cli/tests/capability_exec.rs", "sha256": "1a19972c41c0c1183f549cba9601c6e5742208c6888cff223a8de7e587f26c53", "testName": "inventory_rg_exec_emits_one_result_and_stable_real_rg_artifact" } }
   ],
   "economics": { "benefit": { "metric": "registered production inventory operations with recomputable invocation trace", "value": 2, "unit": "operations" }, "cost": { "metric": "unclosed executable lifecycle gaps", "value": 4, "unit": "gaps" }, "benefitEvidence": { "evidenceIds": ["local:r09:operation-trace", "local:r09:scope-exclusion-conformance"], "checkedAt": 1783900800, "expiresAt": 1791676800 }, "costEvidence": { "evidenceIds": ["gap:rg:package-provenance", "gap:rg:local-license-copy", "gap:rg:replacement-command-drill", "gap:rg:latency-p50-p95-measurement"], "checkedAt": 1783900800, "expiresAt": 1791676800 } },
